ProjectManager refactor WIP Clearing data after tests is problematic

+++ src/main/java/es/ucm/fdi/business/workspace/project/ProjectManagerAS.java
@@ -13,8 +13,13 @@
  */
 package es.ucm.fdi.business.workspace.project;
 
-import es.ucm.fdi.integration.project.ProjectDAO;
+import java.security.AccessControlException;
+
+import es.ucm.fdi.business.users.SessionDTO;
+import es.ucm.fdi.business.users.UserManagerAS;
+import es.ucm.fdi.integration.project.ProjectDAO;	
 import es.ucm.fdi.integration.project.ProjectDAOSQLImp;
+import es.ucm.fdi.integration.users.UserDAOHashTableImp;
 
 /**
  * Application service to manage the traffic of projects.
@@ -23,7 +28,8 @@ import es.ucm.fdi.integration.project.ProjectDAOSQLImp;
  */
 public class ProjectManagerAS {
 
-
+	private final UserManagerAS userMan = UserManagerAS
+            .getManager(new UserDAOHashTableImp());
     private final ProjectDAO dao;
     private final String user;
 
@@ -59,17 +65,21 @@ public class ProjectManagerAS {
      *
      * @param proj The project to add.
      */
-    public void newProject(ProjectDTO proj) {
-        if (validateProjectDetails(proj)) {
-            if (dao.findProject(proj.getID()) == null) {
-                dao.addProject(proj);
-            } else {
-                throw new IllegalArgumentException("Project " + proj.getID()
-                        + " already exists");
-            }
-        } else {
-            throw new IllegalArgumentException("Invalid project details");
-        }
+    public void newProject(ProjectDTO proj, SessionDTO session) {
+    	if (userMan.authenticate(user, session)) {
+	        if (validateProjectDetails(proj)) {
+	            if (dao.findProject(proj.getID()) == null) {
+	                dao.addProject(proj);
+	            } else {
+	                throw new IllegalArgumentException("Project " + proj.getID()
+	                        + " already exists");
+	            }
+	        } else {
+	            throw new IllegalArgumentException("Invalid project details");
+	        }
+    	} else {
+    		throw new AccessControlException("Invalid session");
+    	}
     }
 
     /**
@@ -77,12 +87,16 @@ public class ProjectManagerAS {
      *
      * @param id The id of the project to be deleted.
      */
-    public void removeProject(String id) {
-        if (dao.findProject(id) != null) {
-            dao.removeProject(id);
-        } else {
-            throw new IllegalArgumentException("Project " + id + " does not exist");
-        }
+    public void removeProject(String id, SessionDTO session) {
+    	if (userMan.authenticate(user, session)) {
+	        if (dao.findProject(id) != null) {
+	            dao.removeProject(id);
+	        } else {
+	            throw new IllegalArgumentException("Project " + id + " does not exist");
+	        }
+    	} else {
+    		throw new AccessControlException("Invalid session");
+    	}
     }
 
     /**
@@ -91,12 +105,16 @@ public class ProjectManagerAS {
      * @param id The id of the project to be opened.
      * @return the project.
      */
-    public ProjectDTO openProject(String id) {
-        ProjectDTO proj = dao.findProject(id);
-        if (proj == null) {
-            throw new IllegalArgumentException("Project " + id + " does not exist");
-        }
-        return proj;
+    public ProjectDTO openProject(String id, SessionDTO session) {
+    	if (userMan.authenticate(user, session)) {
+	        ProjectDTO proj = dao.findProject(id);
+	        if (proj == null) {
+	            throw new IllegalArgumentException("Project " + id + " does not exist");
+	        }
+	        return proj;
+    	} else {
+    		throw new AccessControlException("Invalid session");
+    	}
     }
 
     /**
@@ -104,12 +122,16 @@ public class ProjectManagerAS {
      *
      * @param proj Project
      */
-    public void saveChanges(ProjectDTO proj) {
-        if (dao.containsProject(proj.getID())) {
-            dao.modifyProject(proj);
-        } else {
-            newProject(proj);
-        }
+    public void modifyProject(ProjectDTO proj, SessionDTO session) {
+    	if (userMan.authenticate(user, session)) {
+	        if (dao.containsProject(proj.getID())) {
+	            dao.modifyProject(proj);
+	        } else {
+	            newProject(proj, session);
+	        }
+    	} else {
+    		throw new AccessControlException("Invalid session");
+    	}
     }
 
     /**
